Display latest articles in an infinite scrolling list

+++ api/gql/schema.js
@@ -14,7 +14,7 @@ module.exports = buildSchema(`
   type Article {
     id: ID!
     title: String!
-    created: Int!
+    created: Float!
     next: String
     prev: String
     content: [Node]
